requirements.txt
langchain-core
langchain-community
pydantic
pandas
openpyxl
olefile
zlib-state
beautifulsoup4
unstructured
pdfplumber
tabulate
neo4j
python-dotenv
langchain-ollama
langchain-huggingface
sentence-transformers
arize-phoenix[evals]
ragas
openinference-instrumentation-langchain
chainlit
